# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contacts API",
    "description": "API documentation for managing contacts (CSE341 - W02 Project: Contacts Part 2). Armenta",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "Contacts",
      "description": "Endpoints for managing contacts"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/contacts/": {
      "get": {
        "summary": "Get all contacts",
        "responses": {
          "200": {
            "description": "List of all contacts",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Contact"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Create a new contact",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Contact data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Contact created successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "summary": "Get a contact by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Contact ID"
          }
        ],
        "responses": {
          "200": { "description": "Contact found" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Contact not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "summary": "Update a contact by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Contact ID"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated contact data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "204": { "description": "Contact updated successfully" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "summary": "Delete a contact by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Contact ID"
          }
        ],
        "responses": {
          "204": { "description": "Contact deleted successfully" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "type": "object",
      "properties": {
        "firstName": { "type": "string", "example": "Ana" },
        "lastName": { "type": "string", "example": "Pérez" },
        "email": { "type": "string", "example": "ana@example.com" },
        "favoriteColor": { "type": "string", "example": "Blue" },
        "birthday": { "type": "string", "example": "1998-04-12" }
      }
    }
  }
}